chore: restructure and combine dash/chart components
