+++ packages/frontend/src/components/Home/LandingPanel/index.tsx
@@ -1,11 +1,7 @@
 import { FC } from 'react';
-import SpaceBrowser from '../../Explorer/SpaceBrowser';
-import LatestDashboards from '../LatestDashboards';
-import LatestSavedCharts from '../LatestSavedCharts';
 import {
     Intro,
     LandingHeaderWrapper,
-    LandingPanelWrapper,
     StyledLinkButton,
     Title,
     WelcomeText,
@@ -14,42 +10,33 @@ import {
 interface Props {
     userName: string | undefined;
     projectUuid: string;
-    hasSavedChart: boolean;
 }
 
-const LandingPanel: FC<Props> = ({ userName, projectUuid, hasSavedChart }) => {
+const LandingPanel: FC<Props> = ({ userName, projectUuid }) => {
     return (
-        <LandingPanelWrapper>
-            <LandingHeaderWrapper>
-                <WelcomeText>
-                    <Title>
-                        {`Welcome${
-                            userName ? ', ' + userName : ' to Lightdash'
-                        }! ⚡`}
-                    </Title>
+        <LandingHeaderWrapper>
+            <WelcomeText>
+                <Title>
+                    {`Welcome${
+                        userName ? ', ' + userName : ' to Lightdash'
+                    }! ⚡`}
+                </Title>
 
-                    <Intro>
-                        Run a query to ask a business question or browse your
-                        data below:
-                    </Intro>
-                </WelcomeText>
+                <Intro>
+                    Run a query to ask a business question or browse your data
+                    below:
+                </Intro>
+            </WelcomeText>
 
-                <StyledLinkButton
-                    large
-                    href={`/projects/${projectUuid}/tables`}
-                    intent="primary"
-                    icon="series-search"
-                >
-                    Run a query
-                </StyledLinkButton>
-            </LandingHeaderWrapper>
-
-            <SpaceBrowser projectUuid={projectUuid} />
-
-            {hasSavedChart && <LatestDashboards projectUuid={projectUuid} />}
-
-            <LatestSavedCharts projectUuid={projectUuid} />
-        </LandingPanelWrapper>
+            <StyledLinkButton
+                large
+                href={`/projects/${projectUuid}/tables`}
+                intent="primary"
+                icon="series-search"
+            >
+                Run a query
+            </StyledLinkButton>
+        </LandingHeaderWrapper>
     );
 };
