+++ packages/frontend/src/pages/Home.tsx
@@ -5,9 +5,11 @@ import { useUnmount } from 'react-use';
 import ErrorState from '../components/common/ErrorState';
 import Page from '../components/common/Page/Page';
 import { PageContentWrapper } from '../components/common/Page/Page.styles';
+import SpaceBrowser from '../components/Explorer/SpaceBrowser';
 import ForbiddenPanel from '../components/ForbiddenPanel';
 import LandingPanel from '../components/Home/LandingPanel';
 import OnboardingPanel from '../components/Home/OnboardingPanel/index';
+import RecentlyUpdatedPanel from '../components/Home/RecentlyUpdatedPanel';
 import {
     useOnboardingStatus,
     useProjectSavedChartStatus,
@@ -60,11 +62,14 @@ const Home: FC = () => {
                     />
                 ) : (
                     <LandingPanel
-                        hasSavedChart={!!savedChartStatus.data}
                         userName={user.data?.firstName}
                         projectUuid={project.data.projectUuid}
                     />
                 )}
+
+                <SpaceBrowser projectUuid={project.data.projectUuid} />
+
+                <RecentlyUpdatedPanel projectUuid={project.data.projectUuid} />
             </PageContentWrapper>
         </Page>
     );
